requirements.txt
flask
flask_sqlalchemy
python-dotenv
openai
faiss-cpu
